validate outfilename
