+++ packages/cli/src/user-passed-output-location.ts
@@ -10,13 +10,3 @@ export const getUserPassedOutputLocation = () => {
 	const filename = parsedCli._[3];
 	return filename;
 };
-
-export const getUserPassedFileExtension = () => {
-	const filename = getUserPassedOutputLocation();
-	const filenameArr = filename.split('.');
-
-	const hasExtension = filenameArr.length >= 2;
-	const filenameArrLength = filenameArr.length;
-	const extension = hasExtension ? filenameArr[filenameArrLength - 1] : null;
-	return extension;
-};
